+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidyr (development version)
 
+* `fill()` gains a `.by` argument as an alternative to `dplyr::group_by()` for
+  applying the fill per group, similar to `nest(.by =)` and
+  `dplyr::mutate(.by =)` (@olivroy, #1439).
+
 * `unchop()` produces a more helpful error message when columns cannot be cast
   to `ptype` (@mgirlich, #1477).
 
@@ -14,8 +18,6 @@
   inputs (@catalamarti, #1510).
 
 * tidyr now requires dplyr >=1.1.0 (#1568, @catalamarti).
-
-* `fill()` gains `.by` similar to how `mutate()` works (@olivroy, #1439).
 
 # tidyr 1.3.1